Allow passing clone function to saveNode
Since 'structuredClone' does not work for Proxy objects,
and it is still necessary to use while the 'content' instance
variable could be a reactive object (i.e. proxy object), we
allow passing the clone function to the saveNode function. This
means downstream apps like shacl-vue can pass a clonefunction
that deals with the proxy object cloning accordingly (in that
case, first running the object through vue's 'toRaw')